requirements.txt
pyserial
paho-mqtt
flask
dataset
mysqlclient
sqlalchemy_utils